Add files via upload
CCTV cameras are installed all around the world in each country and each city but still the crime rate is really increasing rapidly as countries do not have the manpower to monitor the all the events and places, so authorities can implement an intelligent crime detection system using deep learning. Suggesting utilizing both regular and abnormal videos to learn anomalies . We Will be using a crime dataset . It includes 10,000 lengthy and uncut real-world surveillance footage, as well as 3000 actual abnormalities such as Assault,accidents, burglaries, robberies, and other routine events. 

The goal of this project is to be able to detect the crime happening without involving any human in the process and raise an alarm to the authorized user to fasten the process.